+++ src/state/modules/actions.js
@@ -1,26 +1,25 @@
 import axios from '@/services/axios'
 
 export const actions = {
-    doAction(_, { id, data, type }){
-        return  axios
+    doAction(_, { id, type, action }) {
+        return axios
             .post(
-                `${type}/${id}/actions`, 
-                data
+                `${type}/${id}/${action}`
             )
-            .then(response=>{
+            .then(response => {
                 return response.data;
             })
     },
-    doLike(_, { id }){
-        return actions.doAction(_, {type: "files", id: id, data: { type: "like" }})
+    doLike(_, { id }) {
+        return actions.doAction(_, { type: "files", id: id, action: "like" })
     },
-    doUnLike(_, { id }){
-        return actions.doAction(_, {type: "files", id: id, data: { type: "unlike" }})
+    doUnLike(_, { id }) {
+        return actions.doAction(_, { type: "files", id: id, action: "unlike" })
     },
-    doFollow(_, { id }){
-        return actions.doAction(_, {type: "users", id: id, data : { type: "follow" }})
+    doFollow(_, { id }) {
+        return actions.doAction(_, { type: "users", id: id, action: "follow" })
     },
-    doUnFollow(_, { id }){
-        return actions.doAction(_, {type: "users", id: id, data : { type: "unfollow" }})
+    doUnFollow(_, { id }) {
+        return actions.doAction(_, { type: "users", id: id, action: "unfollow" })
     }
 }