+++ src/modules/profile/pages/Profile.vue
@@ -182,7 +182,17 @@ export default {
       paginator.nextPage().then((data) => this.appendToTab(tab, data));
     },
     appendToTab(tab, data = []) {
-      this.data[tab] = [...(this.data[tab]||[]), ...(data.items||[])];
+      this.data[tab] = this.data[tab] || [];
+
+      data.items.forEach((item) => {
+        console.log(item);
+        if (
+          !item.id ||
+          this.data[tab].filter((_item) => _item.id == item.id).length === 0
+        ) {
+          this.data[tab].push(item);
+        }
+      });
     },
   },
   async beforeMount() {